H2M/requirements.txt
# Core API
fastapi
uvicorn[standard]
pydantic
websockets

# In-Memory Data Grid for Conversation History
pyignite

# Configuration
pyyaml

# Service Clients
httpx

# Utilities
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi

# Local Shared Libraries (install in editable mode from project root)
# These dependencies are on local packages.
# They are listed here for reference.
# -e ./shared/q_auth_parser
# -e ./shared/q_vectorstore_client
# -e ./shared/q_pulse_client 